requirements.txt
keras_vggface==0.6
mtcnn==0.1.0
numpy==1.18.5
opencv-python
opencv-python-headless
Pillow==9.1.1
pip==22.1.2
sklearn
streamlit==1.10.0
tensorflow==2.3.1
tqdm==4.64.0
keras==2.4.3
keras_applications==1.0.8